Introduced HTTP_HEADERS (http_headers.txt) and PREFIX_HTTP_HEADERS. Default is to NOT prefix headers.
Handler for http is now processing all headers before sending to process_request(),
and process_request() is now being run in main shell, not subshell.

Request body appears to be working for http handler but not when headers are prefixed,
because Haserl expects 'CONTENT_TYPE' and possibly 'CONTENT_LENGTH' with no prefix.